refactor!: Remove unstyled button
According to #45 we're using button to compose other components, and to ease that implementation we're using this unstyled variant.

IMHO, this isn't what variants should be used for. It's much better to have a ordinary Javascript object which has the values to reset a button than using the variant concept to solve this. My reasoning is that an unstyled button is not a variant of a button. Essentially, a button variant is typically `primary`, `outlined` which have two use-cases in a design.